(PC-17816) refactor(profile): move AlreadyChangedEmailDisclaimer and ChangeEmailDisclaimer from pages to components
